animation for main page added & renamed components
