--- index.html
@@ -1,79 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.14.0/css/all.min.css"
-    integrity="sha512-1PKOgIY59xJ8Co8+NE6FZ+LOAZKjy+KY8iq0G4B3CyeY6wYHN3yt9PW0XpSriVlkMXe40PTKnXrLnZ9+fkDaog=="
-    crossorigin="anonymous" />
-  <link rel="stylesheet" href="./style.css" />
-  <title>Sąsiedzi</title>
-</head>
-
-<body>
-  <div class="container">
-    <div class="forms-container">
-      <div class="signin-signup">
-
-        <!-- SIGNING IN -->
-        <form action="" class="sign-in-form">
-          <h2 class="title">Sign in</h2>
-          <div class="input-field">
-            <i class="fas fa-user"></i>
-            <input type="text" placeholder="Username">
-          </div>
-          <div class="input-field">
-            <i class="fas fa-lock"></i>
-            <input type="password" placeholder="Password">
-          </div>
-          <input type="submit" value="Login" class="btn solid">
-
-        </form>
-
-        <!-- SIGNING UP -->
-        <form action="" class="sign-up-form">
-          <h2 class="title">Sign up</h2>
-          <div class="input-field">
-            <i class="fas fa-user"></i>
-            <input type="text" placeholder="Username">
-          </div>
-          <div class="input-field">
-            <i class="fas fa-envelope"></i>
-            <input type="text" placeholder="Email">
-          </div>
-          <div class="input-field">
-            <i class="fas fa-lock"></i>
-            <input type="password" placeholder="Password">
-          </div>
-          <input type="submit" value="Sign up" class="btn solid">
-          </div>
-        </form>
-    </div>
-
-    <div class="pannels-container">
-      <div class="panel left-panel">
-        <div class="content">
-          <h3>New here?</h3>
-          <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolorem dolorum expedita?</p>
-          <button class="btn transparent" id="sign-up-btn">Sign up</button>
-        </div>
-        <img src="./img/login.svg" class="image" alt=""></img>
-      </div>
-
-      <div class="panel right-panel">
-        <div class="content">
-          <h3>One of us?</h3>
-          <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolorem dolorum expedita?</p>
-          <button class="btn transparent" id="sign-in-btn">Sign in</button>
-        </div>
-        <img src="./img/register.svg" class="image" alt=""></img>
-
-      </div>
-    </div>
-  </div>
-
-  <script src="./srcipt.js"></script>
-</body>
-
-</html>
