+++ neighbors/package.json
@@ -10,6 +10,7 @@
     "@types/node": "^16.11.29",
     "@types/react": "^18.0.6",
     "firebase": "^9.9.4",
+    "prop-types": "^15.8.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-router-dom": "^6.3.0",
