+++ neighbors/src/components/containers/container.jsx
@@ -1,12 +1,20 @@
 import '../mainLayout.css';
-import { FormsContainer } from './formsContainer';
-import { PannelsContainer } from './pannelsContainer';
+import { useRef } from 'react';
+
+import { FormsContainer } from './FormsContainer';
+import { PannelsContainer } from './PannelsContainer';
 
 export function Container() {
+  const containerRef = useRef(null);
+  const onButtonClick = () => {
+    containerRef.current;
+
+    containerRef.current.classList.toggle('sign-up-mode');
+  };
   return (
-    <div className="container">
-      <FormsContainer></FormsContainer>
-      <PannelsContainer></PannelsContainer>
+    <div className="container" ref={containerRef}>
+      <FormsContainer />
+      <PannelsContainer onClick={onButtonClick} />
     </div>
   );
 }
