--- neighbors/src/components/loginImage.jsx
@@ -1,10 +1,0 @@
-import LogImg from '../img/login.svg';
-import './mainLayout.css';
-
-export const LoginImage = () => {
-  return (
-    <>
-      <img src={LogImg} className="image" alt="login" />
-    </>
-  );
-};
