--- neighbors/src/components/registerImage.jsx
@@ -1,10 +1,0 @@
-import RegImg from '../img/register.svg';
-import './mainLayout.css';
-
-export const RegisterImage = () => {
-  return (
-    <>
-      <img src={RegImg} className="image" alt="register" />
-    </>
-  );
-};
