--- neighbors/src/components/srcipt.jsx
@@ -1,10 +1,0 @@
-const signInBtn = document.querySelector('#sign-in-btn');
-const signUpBtn = document.querySelector('#sign-up-btn');
-const container = document.querySelector('.container');
-
-signUpBtn.addEventListener('click', () => {
-  container.classList.add('sign-up-mode');
-});
-signInBtn.addEventListener('click', () => {
-  container.classList.remove('sign-up-mode');
-});
